requirements.txt
beautifulsoup4
bibtexparser
lxml
pybtex
python-dateutil
requests
